[Bug] 잔디 스트릭 관련 오류 수정

+++ src/main/java/develop/grassserver/grass/presentation/dto/MemberStreakResponse.java
@@ -3,6 +3,6 @@ package develop.grassserver.grass.presentation.dto;
 public record MemberStreakResponse(
         int currentStreak,
         int maxStreak,
-        int totalStudyTime
+        long totalStudyTime
 ) {
 }
